Fix: 2 map files for 11-2
